RavenDB-12204: change counter value to double when exporting a DB, caused incorrect values on very big numbers.

+++ src/Raven.Server/Smuggler/Documents/StreamDestination.cs
@@ -422,7 +422,7 @@ namespace Raven.Server.Smuggler.Documents
                 Writer.WriteComma();
 
                 Writer.WritePropertyName(nameof(DocumentItem.CounterItem.Value));
-                Writer.WriteDouble(counterDetail.TotalValue);
+                Writer.WriteInteger(counterDetail.TotalValue);
                 Writer.WriteComma();
 
                 Writer.WritePropertyName(nameof(DocumentItem.CounterItem.ChangeVector));
